nfc: Declare `status` in cold_reset struct as signed 16-bit integer
nfc_ese_pwr() can assign -EIO to `status`, but it is unsigned. So, it
cannot take negative integers into account. Switch to signed 16-bit
integer instead.

(cherry picked from commit d97d7ef0a53cbf5d672a30da2759ec9a0d2b8b7a)

+++ drivers/nfc/qti/nfc_common.h
@@ -205,7 +205,7 @@ struct platform_ldo {
 struct cold_reset {
 	wait_queue_head_t read_wq;
 	bool rsp_pending;
-	uint8_t status;
+	s16 status;
 	/* Is NFC enabled from UI */
 	bool is_nfc_enabled;
 };
